dev: refactored FireStore services
