+++ src/services/authentication/authentication.ts
@@ -48,7 +48,7 @@ export const logIn = createAsyncThunk(
 
 export const logOut = createAsyncThunk(
   'auth/logOut',
-  async (_, { rejectWithValue, dispatch }) => {
+  async (_, { rejectWithValue }) => {
     const auth = getAuth()
 
     try {
