+++ vite.config.ts
@@ -19,6 +19,7 @@ export default defineConfig({
       assets: '/src/assets',
       components: '/src/components',
       services: '/src/services',
+      api: '/src/api',
       uikit: '/src/UIkit',
       utils: '/src/utils',
       pages: '/src/pages',
